Extract API functions out per endpoint
