+++ lib/endpoints.rb
@@ -1,0 +1,28 @@
+require 'endpoints/assets'
+require 'endpoints/blobstore'
+require 'endpoints/components'
+require 'endpoints/content-selectors'
+require 'endpoints/email'
+require 'endpoints/formats'
+require 'endpoints/iq'
+require 'endpoints/ldap'
+require 'endpoints/license'
+require 'endpoints/lifecycle'
+require 'endpoints/privileges'
+require 'endpoints/read-only'
+require 'endpoints/realms'
+require 'endpoints/repositories'
+require 'endpoints/roles'
+require 'endpoints/routing-rules'
+require 'endpoints/saml'
+require 'endpoints/script'
+require 'endpoints/search'
+require 'endpoints/ssl'
+require 'endpoints/staging'
+require 'endpoints/status'
+require 'endpoints/support'
+require 'endpoints/tags'
+require 'endpoints/tasks'
+require 'endpoints/user-sources'
+require 'endpoints/user-tokens'
+require 'endpoints/users'
