+++ lib/endpoints/blobstore.rb
@@ -1,0 +1,17 @@
+module NexusAPI
+  class API
+    # GET /service/rest/beta/blobstores
+    # DELETE /service/rest/beta/blobstores/{name}
+    # POST /service/rest/beta/blobstores/file
+    # GET /service/rest/beta/blobstores/file/{name}
+    # PUT /service/rest/beta/blobstores/file/{name}
+    # POST /service/rest/beta/blobstores/group
+    # GET /service/rest/beta/blobstores/group/{name}
+    # PUT /service/rest/beta/blobstores/group/{name}
+    # POST /service/rest/beta/blobstores/group/promote/{name}
+    # POST /service/rest/beta/blobstores/s3
+    # GET /service/rest/beta/blobstores/s3/{name}
+    # PUT /service/rest/beta/blobstores/s3/{name}
+    # GET /service/rest/v1/blobstores/{id}/quota-status
+  end
+end
