+++ lib/endpoints/content-selectors.rb
@@ -1,0 +1,9 @@
+module NexusAPI
+  class API
+    # GET /service/rest/beta/security/content-selectors
+    # POST /service/rest/beta/security/content-selectors
+    # GET /service/rest/beta/security/content-selectors/{name}
+    # PUT /service/rest/beta/security/content-selectors/{name}
+    # DELETE /service/rest/beta/security/content-selectors/{name}
+  end
+end
