+++ lib/endpoints/email.rb
@@ -1,0 +1,8 @@
+module NexusAPI
+  class API
+    # GET /service/rest/beta/email
+    # PUT /service/rest/beta/email
+    # DELETE /service/rest/beta/email
+    # POST /service/rest/beta/email/verify
+  end
+end
