+++ lib/endpoints/formats.rb
@@ -1,0 +1,6 @@
+module NexusAPI
+  class API
+    # GET /service/rest/v1/formats/{format}/upload-specs
+    # GET /service/rest/v1/formats/upload-specs
+  end
+end
