+++ lib/endpoints/iq.rb
@@ -1,0 +1,9 @@
+module NexusAPI
+  class API
+    # GET /service/rest/beta/iq
+    # PUT /service/rest/beta/iq
+    # POST /service/rest/beta/iq/disable
+    # POST /service/rest/beta/iq/enable
+    # POST /service/rest/beta/iq/verify-connection
+  end
+end
