+++ lib/endpoints/ldap.rb
@@ -1,0 +1,10 @@
+module NexusAPI
+  class API
+    # GET /service/rest/beta/security/ldap
+    # POST /service/rest/beta/security/ldap
+    # GET /service/rest/beta/security/ldap/{name}
+    # PUT /service/rest/beta/security/ldap/{name}
+    # DELETE /service/rest/beta/security/ldap/{name}
+    # POST /service/rest/beta/security/ldap/change-order
+  end
+end
