+++ lib/endpoints/license.rb
@@ -1,0 +1,7 @@
+module NexusAPI
+  class API
+    # GET /service/rest/beta/system/license
+    # POST /service/rest/beta/system/license
+    # DELETE /service/rest/beta/system/license
+  end
+end
