+++ lib/endpoints/lifecycle.rb
@@ -1,0 +1,7 @@
+module NexusAPI
+  class API
+    # PUT /service/rest/v1/lifecycle/bounce
+    # GET /service/rest/v1/lifecycle/phase
+    # PUT /service/rest/v1/lifecycle/phase
+  end
+end
