+++ lib/endpoints/read-only.rb
@@ -1,0 +1,8 @@
+module NexusAPI
+  class API
+    # GET /service/rest/v1/read-only
+    # POST /service/rest/v1/read-only/force-release
+    # POST /service/rest/v1/read-only/freeze
+    # POST /service/rest/v1/read-only/release
+  end
+end
