+++ lib/endpoints/realms.rb
@@ -1,0 +1,7 @@
+module NexusAPI
+  class API
+    # GET /service/rest/beta/security/realms/active
+    # PUT /service/rest/beta/security/realms/active
+    # GET /service/rest/beta/security/realms/available
+  end
+end
