+++ lib/endpoints/roles.rb
@@ -1,0 +1,9 @@
+module NexusAPI
+  class API
+    # GET /service/rest/beta/security/roles
+    # POST /service/rest/beta/security/roles
+    # GET /service/rest/beta/security/roles/{id}
+    # PUT /service/rest/beta/security/roles/{id}
+    # DELETE /service/rest/beta/security/roles/{id}
+  end
+end
