+++ lib/endpoints/routing-rules.rb
@@ -1,0 +1,9 @@
+module NexusAPI
+  class API
+    # GET /service/rest/beta/routing-rules
+    # POST /service/rest/beta/routing-rules
+    # GET /service/rest/beta/routing-rules/{name}
+    # PUT /service/rest/beta/routing-rules/{name}
+    # DELETE /service/rest/beta/routing-rules/{name}
+  end
+end
