+++ lib/endpoints/saml.rb
@@ -1,0 +1,9 @@
+module NexusAPI
+  class API
+    # GET /service/rest/v1/security/saml
+    # PUT /service/rest/v1/security/saml
+    # DELETE /service/rest/v1/security/saml
+    # GET /service/rest/v1/security/saml/metadata
+    # GET /service/rest/v1/security/saml/pem
+  end
+end
