+++ lib/endpoints/ssl.rb
@@ -1,0 +1,8 @@
+module NexusAPI
+  class API
+    # GET /service/rest/beta/security/ssl
+    # GET /service/rest/beta/security/ssl/truststore
+    # POST /service/rest/beta/security/ssl/truststore
+    # DELETE /service/rest/beta/security/ssl/truststore/{id}
+  end
+end
