+++ lib/endpoints/staging.rb
@@ -1,0 +1,6 @@
+module NexusAPI
+  class API
+    # POST /service/rest/v1/staging/delete
+    # POST /service/rest/v1/staging/move/{destination}
+  end
+end
