+++ lib/endpoints/support.rb
@@ -1,0 +1,5 @@
+module NexusAPI
+  class API
+    # POST /service/rest/v1/support/supportzip
+  end
+end
