+++ lib/endpoints/tasks.rb
@@ -1,0 +1,8 @@
+module NexusAPI
+  class API
+    # GET /service/rest/v1/tasks
+    # GET /service/rest/v1/tasks/{id}
+    # POST /service/rest/v1/tasks/{id}/run
+    # POST /service/rest/v1/tasks/{id}/stop
+  end
+end
