+++ lib/endpoints/user-sources.rb
@@ -1,0 +1,5 @@
+module NexusAPI
+  class API
+    # GET /service/rest/beta/security/user-sources
+  end
+end
