+++ lib/endpoints/user-tokens.rb
@@ -1,0 +1,5 @@
+module NexusAPI
+  class API
+    # DELETE /service/rest/beta/security/user-tokens
+  end
+end
