+++ lib/endpoints/users.rb
@@ -1,0 +1,10 @@
+module NexusAPI
+  class API
+    # GET /service/rest/beta/security/users
+    # POST /service/rest/beta/security/users
+    # PUT /service/rest/beta/security/users/{userId}
+    # DELETE /service/rest/beta/security/users/{userId}
+    # PUT /service/rest/beta/security/users/{userId}/change-password
+    # DELETE /service/rest/beta/security/users/{userId}/user-token
+  end
+end
